moving XmlMetadataTemplate to doi package
