+++ src/main/java/edu/harvard/iq/dataverse/pidproviders/AbstractPidProvider.java
@@ -456,43 +456,6 @@ public abstract class AbstractPidProvider implements PidProvider {
         return identifier;
     }
 
-    public String getMetadataFromDvObject(String identifier, Map<String, String> metadata, DvObject dvObject) {
-
-        Dataset dataset = null;
-
-        if (dvObject instanceof Dataset) {
-            dataset = (Dataset) dvObject;
-        } else {
-            dataset = (Dataset) dvObject.getOwner();
-        }
-
-        XmlMetadataTemplate metadataTemplate = new XmlMetadataTemplate();
-        metadataTemplate.setIdentifier(identifier.substring(identifier.indexOf(':') + 1));
-        metadataTemplate.setCreators(Arrays.asList(metadata.get("datacite.creator").split("; ")));
-        metadataTemplate.setAuthors(dataset.getLatestVersion().getDatasetAuthors());
-        if (dvObject.isInstanceofDataset()) {
-            metadataTemplate.setDescription(dataset.getLatestVersion().getDescriptionPlainText());
-        }
-        if (dvObject.isInstanceofDataFile()) {
-            DataFile df = (DataFile) dvObject;
-            String fileDescription = df.getDescription();
-            metadataTemplate.setDescription(fileDescription == null ? "" : fileDescription);
-        }
-
-        metadataTemplate.setContacts(dataset.getLatestVersion().getDatasetContacts());
-        metadataTemplate.setProducers(dataset.getLatestVersion().getDatasetProducers());
-        metadataTemplate.setTitle(dvObject.getCurrentName());
-        String producerString = pidProviderService.getProducer();
-        if (producerString.isEmpty() || producerString.equals(DatasetField.NA_VALUE)) {
-            producerString = UNAVAILABLE;
-        }
-        metadataTemplate.setPublisher(producerString);
-        metadataTemplate.setPublisherYear(metadata.get("datacite.publicationyear"));
-
-        String xmlMetadata = metadataTemplate.generateXML(dvObject);
-        logger.log(Level.FINE, "XML to send to DataCite: {0}", xmlMetadata);
-        return xmlMetadata;
-    }
 
     @Override
     public boolean canManagePID() {
