+++ src/main/java/edu/harvard/iq/dataverse/pidproviders/doi/AbstractDOIProvider.java
@@ -1,11 +1,22 @@
 package edu.harvard.iq.dataverse.pidproviders.doi;
 
+import java.util.Arrays;
+import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import edu.harvard.iq.dataverse.DataFile;
+import edu.harvard.iq.dataverse.Dataset;
+import edu.harvard.iq.dataverse.DatasetField;
+import edu.harvard.iq.dataverse.DvObject;
 import edu.harvard.iq.dataverse.GlobalId;
 import edu.harvard.iq.dataverse.pidproviders.AbstractPidProvider;
 import edu.harvard.iq.dataverse.pidproviders.PidProvider;
 
 
 public abstract class AbstractDOIProvider extends AbstractPidProvider {
+
+    private static final Logger logger = Logger.getLogger(AbstractDOIProvider.class.getCanonicalName());
 
     public static final String DOI_PROTOCOL = "doi";
     public static final String DOI_RESOLVER_URL = "https://doi.org/";
@@ -70,4 +81,43 @@ public abstract class AbstractDOIProvider extends AbstractPidProvider {
     public String getProtocol() {
         return DOI_PROTOCOL;
     }
+    
+    public String getMetadataFromDvObject(String identifier, Map<String, String> metadata, DvObject dvObject) {
+
+        Dataset dataset = null;
+
+        if (dvObject instanceof Dataset) {
+            dataset = (Dataset) dvObject;
+        } else {
+            dataset = (Dataset) dvObject.getOwner();
+        }
+
+        XmlMetadataTemplate metadataTemplate = new XmlMetadataTemplate();
+        metadataTemplate.setIdentifier(identifier.substring(identifier.indexOf(':') + 1));
+        metadataTemplate.setCreators(Arrays.asList(metadata.get("datacite.creator").split("; ")));
+        metadataTemplate.setAuthors(dataset.getLatestVersion().getDatasetAuthors());
+        if (dvObject.isInstanceofDataset()) {
+            metadataTemplate.setDescription(dataset.getLatestVersion().getDescriptionPlainText());
+        }
+        if (dvObject.isInstanceofDataFile()) {
+            DataFile df = (DataFile) dvObject;
+            String fileDescription = df.getDescription();
+            metadataTemplate.setDescription(fileDescription == null ? "" : fileDescription);
+        }
+
+        metadataTemplate.setContacts(dataset.getLatestVersion().getDatasetContacts());
+        metadataTemplate.setProducers(dataset.getLatestVersion().getDatasetProducers());
+        metadataTemplate.setTitle(dvObject.getCurrentName());
+        String producerString = pidProviderService.getProducer();
+        if (producerString.isEmpty() || producerString.equals(DatasetField.NA_VALUE)) {
+            producerString = UNAVAILABLE;
+        }
+        metadataTemplate.setPublisher(producerString);
+        metadataTemplate.setPublisherYear(metadata.get("datacite.publicationyear"));
+
+        String xmlMetadata = metadataTemplate.generateXML(dvObject);
+        logger.log(Level.FINE, "XML to send to DataCite: {0}", xmlMetadata);
+        return xmlMetadata;
+    }
+
 }
